added mp4v and search for pesky small offsets.

+++ codec_mp4a.cpp
@@ -20,7 +20,7 @@ Match Codec::mp4aMatch(const unsigned char *start, int maxlength) {
 		Log::debug << "mp4a: Success because of horrible hack.\n";
 	}
 
-	int64_t begin32 = readBE<int32_t>(start);
+	int32_t begin32 = readBE<int32_t>(start);
 	if(start[0] == 0 && begin32 < 1000000)
 		return match;
